moved most of the save functionality to AdminButtons, Save Trail is next
